refactor(testutil): dedicated abci block api for each integration network
