+++ testutil/integration/exrp/upgrade/abci.go
@@ -7,7 +7,8 @@ import (
 
 	storetypes "cosmossdk.io/store/types"
 	abcitypes "github.com/cometbft/cometbft/abci/types"
-	exrpcommon "github.com/xrplevm/node/v5/testutil/integration/exrp/common"
+	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
+	cmttypes "github.com/cometbft/cometbft/types"
 )
 
 // NextBlock is a private helper function that runs the EndBlocker logic, commits the changes,
@@ -43,7 +44,7 @@ func (n *UpgradeIntegrationNetwork) finalizeBlockAndCommit(duration time.Duratio
 	header.Time = newBlockTime
 
 	// FinalizeBlock to run endBlock, deliverTx & beginBlock logic
-	req := exrpcommon.BuildFinalizeBlockReq(header, n.valSet.Validators, nil, nil, txBytes...)
+	req := BuildFinalizeBlockReq(header, n.valSet.Validators, txBytes...)
 
 	res, err := n.app.FinalizeBlock(req)
 	if err != nil {
@@ -66,4 +67,35 @@ func (n *UpgradeIntegrationNetwork) finalizeBlockAndCommit(duration time.Duratio
 	_, err = n.app.Commit()
 
 	return res, err
+}
+
+// buildFinalizeBlockReq is a helper function to build
+// properly the FinalizeBlock request
+func BuildFinalizeBlockReq(header cmtproto.Header, validators []*cmttypes.Validator, txs ...[]byte) *abcitypes.RequestFinalizeBlock {
+	// add validator's commit info to allocate corresponding tokens to validators
+	ci := GetCommitInfo(validators)
+	return &abcitypes.RequestFinalizeBlock{
+		Misbehavior:        nil,
+		Height:             header.Height,
+		DecidedLastCommit:  ci,
+		Hash:               header.AppHash,
+		NextValidatorsHash: header.ValidatorsHash,
+		ProposerAddress:    header.ProposerAddress,
+		Time:               header.Time,
+		Txs:                txs,
+	}
+}
+
+func GetCommitInfo(validators []*cmttypes.Validator) abcitypes.CommitInfo {
+	voteInfos := make([]abcitypes.VoteInfo, len(validators))
+	for i, val := range validators {
+		voteInfos[i] = abcitypes.VoteInfo{
+			Validator: abcitypes.Validator{
+				Address: val.Address,
+				Power:   val.VotingPower,
+			},
+			BlockIdFlag: cmtproto.BlockIDFlagCommit,
+		}
+	}
+	return abcitypes.CommitInfo{Votes: voteInfos}
 }
